requirements.txt
requests==2.32.3
psycopg2-binary==2.9.9
pandas==2.2.2
apache-airflow==2.10.0  # Updated to the latest version
# dbt-core==1.8.2
# dbt-postgres==1.8.2
great_expectations==0.18.1
# If you intend to use the Postgres provider specifically, you can list it.
# However, installing it as an extra with apache-airflow[postgres] is often simpler
# as the constraints file will ensure the correct provider version.
# apache-airflow-providers-postgres==4.0.0 # This version might be outdated for Airflow 2.10.0, the constraints file will handle it better.